lint: add almost all annotations to pass `mypy --strict`
Added type aliases bytearray2, bytearray4, etc to signal that the
bytearrays passed must be 2 bytes, 4 bytes, etc.

In adafruit_tinylora_encryption.py, there's a "state" matrix for the AES
implementation there. It was mixing str and byte/int, which apparently
works fine on circuitpython but fails on cpython. Change the state
matrix to a bytearray, as was likely intended.

Fixes #50.